allow configuration values to persist between test tables in standalone mode
If DbFit is in control of the test page with Flow mode, then RunKettleFixture
test tables cannot be as well. We do not want to have to declare every
single configuration variable each time we run a transformation or job.
So make the config variables static so that they persist between each 
test table.